Lambda 2022-12-26 bi 模块API  接口

+++ erp-model/erp-model-bi/src/main/java/com/erp/model/bi/vo/SalesBaseVO.java
@@ -3,6 +3,7 @@ package com.erp.model.bi.vo;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -25,7 +26,7 @@ public class SalesBaseVO {
     /**
      * 销售额
      */
-    private Double sales;
+    private BigDecimal sales;
 
 
     /**
